Moving CDT implementation to new project.
