+++ LibsAndSamples.sln
@@ -185,6 +185,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CacheExtension", "tests\dev
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Identity.Client.Extensions.Msal", "src\client\Microsoft.Identity.Client.Extensions.Msal\Microsoft.Identity.Client.Extensions.Msal.csproj", "{87679336-95BE-47E4-B42B-8F6860A0B215}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MsalCdtExtension", "MsalCdtExtension\MsalCdtExtension.csproj", "{71FFABC1-A20A-48CC-86DD-4D28541F2359}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug + MobileApps|Any CPU = Debug + MobileApps|Any CPU
@@ -1741,6 +1743,48 @@ Global
 		{87679336-95BE-47E4-B42B-8F6860A0B215}.Release|x64.Build.0 = Release|Any CPU
 		{87679336-95BE-47E4-B42B-8F6860A0B215}.Release|x86.ActiveCfg = Release|Any CPU
 		{87679336-95BE-47E4-B42B-8F6860A0B215}.Release|x86.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|Any CPU.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|Any CPU.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|ARM.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|ARM.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|ARM64.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|ARM64.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|iPhone.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|iPhone.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|x64.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|x64.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|x86.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug + MobileApps|x86.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|ARM.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|ARM.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|x64.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Debug|x86.Build.0 = Debug|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|Any CPU.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|ARM.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|ARM.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|ARM64.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|iPhone.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|x64.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|x64.Build.0 = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|x86.ActiveCfg = Release|Any CPU
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1793,6 +1837,7 @@ Global
 		{74805FE3-2E0D-4EAB-8CFE-A9D46F8D5972} = {34BE693E-3496-45A4-B1D2-D3A0E068EEDB}
 		{92064C48-0136-48CD-AE8D-C6FEDBC7B639} = {74805FE3-2E0D-4EAB-8CFE-A9D46F8D5972}
 		{87679336-95BE-47E4-B42B-8F6860A0B215} = {1A37FD75-94E9-4D6F-953A-0DABBD7B49E9}
+		{71FFABC1-A20A-48CC-86DD-4D28541F2359} = {34BE693E-3496-45A4-B1D2-D3A0E068EEDB}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {020399A9-DC27-4B82-9CAA-EF488665AC27}
