+++ MsalCdtExtension/CdtAuthenticationScheme.cs
@@ -17,15 +17,10 @@ using Microsoft.Identity.Client.Cache.Items;
 using Microsoft.Identity.Client.Internal;
 using Microsoft.Identity.Client.OAuth2;
 using Microsoft.Identity.Client.Utils;
-#if SUPPORTS_SYSTEM_TEXT_JSON
 using JObject = System.Text.Json.Nodes.JsonObject;
 using JToken = System.Text.Json.Nodes.JsonNode;
-#else
-using Microsoft.Identity.Json;
-using Microsoft.Identity.Json.Linq;
-#endif
 
-namespace Microsoft.Identity.Client
+namespace MsalCdtExtension
 {
     //Temporary location
     public class CdtAuthenticationScheme : IAuthenticationScheme
@@ -76,7 +71,7 @@ namespace Microsoft.Identity.Client
         {
             var temp = "eyJrdHkiOiAiUlNBIiwgIm4iOiAiMUNMeDNXRW1NWlQ3el92Szc2ZHBaVVNwX2kyMEEza0Y0OWVpemtCQTBFSTJ4el9pcldTcm9BamJrRTk4dlp3SFM0QVlQV2I5WEd2YTFPYVNMX0RqQTFPTG1nSll4Uk45cU5jd1lKeGhsN3hqaGJlU25RMUMtR1NNS3ZWRzJnaDdQUlhaaU1xVXFuOWt3UzBXa1RoNDhSREMxR0xhTFFfNzZmb0dZMmo0MlNvel9XYnNRemtnVGo0TDVaVTZTWjJ3QTFwMlZ6WFliOVd1M3A4U2VuV3JCTDUzOWhUZjVGelp0b1E0R2IxNzMzVzFmWVFsUkotYUZVMTFfdEc1Umx2Ui1nSWFweHJMWkFKM1NHM28wQ2ZPa2FaejdKT2RETnJHNnE4akF3ZmJOdFJ1eDYzbnJZZ0FHc3VhemlXalZxRnZiclNMX2Mya3dZaDlZUl9uYVJFOG1RPT0iLCAiZSI6ICJBUUFCIn0%3D";
             return new Dictionary<string, string>() {
-                { OAuth2Parameter.TokenType, Constants.BearerAuthHeaderPrefix},
+                { Constants.TokenType, Constants.BearerAuthHeaderPrefix},
                 { CdtRequestConfirmation, Base64UrlHelpers.Encode(_dsReqCnf)}
             };
         }
@@ -84,8 +79,8 @@ namespace Microsoft.Identity.Client
         public void FormatResult(AuthenticationResult authenticationResult)
         {
             var header = new JObject();
-            header[JsonWebTokenConstants.Algorithm] = NoAlgorythmPrefix;
-            header[JsonWebTokenConstants.Type] = CdtTokenType;
+            header[Constants.Algorithm] = NoAlgorythmPrefix;
+            header[Constants.Type] = CdtTokenType;
 
             //TODO: determine what happens if nonce is not present
             authenticationResult.AdditionalResponseParameters.TryGetValue(CdtNonce, out string nonce);
@@ -94,18 +89,6 @@ namespace Microsoft.Identity.Client
             string constraintToken = CreateJWS(JsonHelper.JsonObjectToString(body), JsonHelper.JsonObjectToString(header), false);
             authenticationResult.AccessToken = constraintToken;
         }
-
-        //public string FormatAccessToken(MsalAccessTokenCacheItem msalAccessTokenCacheItem)
-        //{
-        //    var header = new JObject();
-        //    header[JsonWebTokenConstants.Type] = Constants.JasonWebTokenType;
-        //    header[JsonWebTokenConstants.Algorithm] = Constants.NoAlgorythmPrefix;
-
-        //    var body = CreateCdtBody(msalAccessTokenCacheItem);
-
-        //    string constraintToken = CreateJWS(JsonHelper.JsonObjectToString(body), JsonHelper.JsonObjectToString(header), false);
-        //    return constraintToken;
-        //}
 
         private JObject CreateCdtBody(string secret, string nonce)
         {
@@ -142,8 +125,8 @@ namespace Microsoft.Identity.Client
         private JToken CreateCdtConstraintsJwT(string nonce)
         {
             var header = new JObject();
-            header[JsonWebTokenConstants.Algorithm] = _cdtCryptoProvider.CryptographicAlgorithm;
-            header[JsonWebTokenConstants.Type] = JasonWebTokenType;
+            header[Constants.Algorithm] = _cdtCryptoProvider.CryptographicAlgorithm;
+            header[Constants.Type] = JasonWebTokenType;
 
             var body = new JObject
             {
@@ -190,108 +173,4 @@ namespace Microsoft.Identity.Client
             return sb.ToString();
         }
     }
-
-    public static class CdtClaimTypes
-    {
-        #region JSON keys for Http request
-
-        /// <summary>
-        /// Access token with response cnf
-        /// 
-        /// </summary>
-        public const string Ticket = "t";
-
-        /// <summary>
-        /// Constraints specified by the client
-        /// 
-        /// </summary>
-        public const string ConstraintsToken = "c";
-
-        /// <summary>
-        /// Constraints specified by the client
-        /// 
-        /// </summary>
-        public const string Constraints = "constraints";
-
-        /// <summary>
-        /// Non-standard claim representing a nonce that protects against replay attacks.
-        /// </summary>
-        public const string Nonce = "xms_ds_nonce ";
-
-        /// <summary>
-        /// 
-        /// </summary>
-        public const string Type = "typ";
-
-        #endregion
-    }
-
-    //TODO: Add support for ECD keys
-    public class CdtCryptoProvider
-    {
-        private readonly X509Certificate2 _cert;
-
-        public CdtCryptoProvider(X509Certificate2 cert)
-        {
-            _cert = cert ?? throw new ArgumentNullException(nameof(cert));
-
-            RSA provider = _cert.GetRSAPublicKey();
-            RSAParameters publicKeyParams = provider.ExportParameters(false);
-            CannonicalPublicKeyJwk = ComputeCanonicalJwk(publicKeyParams);
-        }
-
-        public byte[] Sign(byte[] payload)
-        {
-            using (RSA key = _cert.GetRSAPrivateKey())
-            {
-                return key.SignData(
-                    payload,
-                    HashAlgorithmName.SHA256,
-                    RSASignaturePadding.Pss);
-            }
-        }
-
-        public string CannonicalPublicKeyJwk { get; }
-
-        public string CryptographicAlgorithm { get => "PS256"; }
-
-        /// <summary>
-        /// Creates the canonical representation of the JWK.  See https://tools.ietf.org/html/rfc7638#section-3
-        /// The number of parameters as well as the lexicographic order is important, as this string will be hashed to get a thumbprint
-        /// </summary>
-        private static string ComputeCanonicalJwk(RSAParameters rsaPublicKey)
-        {
-            return $@"{{""e"":""{Base64UrlHelpers.Encode(rsaPublicKey.Exponent)}"",""kty"":""RSA"",""n"":""{Base64UrlHelpers.Encode(rsaPublicKey.Modulus)}""}}";
-        }
-    }
-
-    /// <summary>
-    /// Delagated Constraint
-    /// </summary>
-    public class ConstraintDict
-    {
-        public Dictionary<string, string> Constraints { get; set; } = new Dictionary<string, string>();
-    }
-
-    public class Constraint
-    {
-        public string Version { get; set; }
-        public string Type { get; set; }
-        public string Action { get; set; }
-        public List<ConstraintTarget> Targets { get; set; }
-        public Dictionary<string, object> AdditionalProperties { get; set; }
-    }
-
-    public class ConstraintTarget
-    {
-        public string Value { get; set; }
-        public string Policy { get; set; }
-        public Dictionary<string, object> AdditionalProperties { get; set; }
-        public ConstraintTarget(string value, string policy)
-        {
-            Value = value;
-            Policy = policy;
-        }
-    }
-
 }
