+++ tests/Microsoft.Identity.Test.Integration.netcore/HeadlessTests/CdtTests.cs
@@ -31,7 +31,7 @@ namespace Microsoft.Identity.Test.Integration.NetCore.HeadlessTests
         }
 
         [TestMethod]
-        [Ignore("Need to wait for ESTS to release feature from test slice.")]
+        //[Ignore("Need to wait for ESTS to release feature from test slice.")]
         public async Task CDT_WithCertIntegrationTest_Async()
         {
             //Client.Constraint constraint = new Client.Constraint();
